Cast xs and y to torch.tensor in __getitem__

+++ vulcanai2/datasets/tabulardataset.py
@@ -61,7 +61,7 @@ class TabularDataset(Dataset):
         xs = self.df.drop(self.labelColumn, axis=1).iloc[[2]].values.tolist()[0]
         xs = torch.tensor(xs, dtype=torch.float)
         y = self.df[[self.labelColumn]].iloc[[idx]].values.tolist()[0][0]
-        y = torch.tensor(y)
+        y = torch.tensor(y, dtype=torch.long)
         return xs, y
 
     def save_dataframe(self, file_path):
